feat: add resource file generation for localization
This commit adds the necessary configuration to
generate resource files for localization. It includes
updating the API.csproj file to embed the Resources.resx
file and generate the corresponding Resources.Designer.cs
file during the build process. This allows for easier
management and access to localized strings within the API
project.